+++ src/components/Blog/BlogCategoryList/BlogCategoryList.stories.tsx
@@ -3,7 +3,7 @@ import { ComponentStory, ComponentMeta } from "@storybook/react";
 import Component from ".";
 
 export default {
-  title: "Lists/BlogCategoryList",
+  title: "Blogs & Webinars/BlogCategoryList",
   component: Component,
 } as ComponentMeta<typeof Component>;
 
@@ -13,6 +13,7 @@ const Template: ComponentStory<typeof Component> = (args) => {
 
 export const BlogCategoryList = Template.bind({});
 BlogCategoryList.args = {
+  page: "blog-index",
   categories: [
     { slug: "oak-updates", title: "Oak updates" },
     { slug: "lesson-planning", title: "Lesson planning" },